requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0
faiss-cpu>=1.7.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
